Add separate rollback parameters
Rather than reusing the values Spec.Update during a the rollback of an
update, use parameters from a separate Spec.Rollback UpdateConfig. This
allows configurations such as a rolling, gradual update, and a fast
parallel rollback to the original version if things go wrong.

If no Rollback message is provided, defaults will apply.

This also adds a Rollback flag to UpdateServiceRequest so that manually
triggered rollbacks (which are currently client-side operations) can be
initiated server-side, and have the correct update settings applied.  We
also need to make sure not to trigger an automatic rollback from a
failed manual rollback.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>