integration: Be choosy about which manager to demote
In rare cases, TestDemoteDownedManager can fail. The scenario involves
choosing the third manager to demote, and stopping it before it has a
chance to write any entries to its WAL. When it starts back up, it
doesn't know of any other members in the cluster (and no other members
will connect to it since it has been removed), so it has no way of
knowing it has been removed.

To avoid this case, we can simply avoid choosing the third manager.
Adding that manager is a synchronization point which guarantees that the
first and second managers will be caught up, and removing either at this
point wouldn't pose the same problem. In practice, we end up removing
the second manager, since the first is the leader.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>