manager: Add a function to check if the raft state is dirty
As one of the requirements for enabling Swarm mode by default, we need
to prevent "swarm init" or "swarm join" commands from unintentionally
overwriting data that was added to Raft previously. We do this by adding
a IsStateDirty method that inspects the contents of the Raft store.
Any data other than a cluster object and object for the local node make
the state dirty.

Note that this function is inherently racy, since new data could be
added while the check is run, or shortly afterwards. However this is
considered acceptable, since the purpose is to check for old data that
the user may accidentally overwrite, not concurrent modifications.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>