Add `ReadOnly` in ContainerSpec for `--read-only`
This fix tries to address the issue raised in docker/29972 where
it was not possible to specify `--read-only` for `docker service create`
and `docker service update`, in order to have the container's root file
system to be read only.

This fix adds `ReadOnly` so that it is possible to specify
`--read-only` for `service update` and `service create`.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>