Add more validation around Replicas parameter
This parameter is declared as a uint64, but if an API user tries to put
a negative number in it and underflows the field, the replicated
orchestrator will panic when it converts the value to an int and tries
to use it as as slice index.

Fix this in the replicated orchestrator by avoiding the conversion to
int, and also add validation in controlapi that the sign bit is unset.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>