Replace gogoproto.customname with plugin that maps ID name
This fix tries to address the issue in 1785 where
gogoproto.customname was repeatedly used manually
to map ID names conforming to golang convention.

This fix address the issue by preprocess the fields and
set `gogoproto.E_Customname` as needed, if name matches
certain pattern:
1. id -> ID
2. id_some -> IDSome
3. some_id -> SomeID
4. some_ids -> SomeIDs

This fix does not change the generated pb.go.

This fix fixes 1785.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>