agent: Leave should wait for the agent to stop accepting assignments
A data race is possible when the agent leaves the cluster. Leave calls
Wait, but Wait may race with starting a new worker when an assignment is
accepted. The existing code signals to the agent's main loop to stop
accepting assignments, but this isn't quite enough because an assignment
may be received at the same time as this signal. Wait for the main loop
to stop accepting assingments before calling Wait.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>