Relocate customnameid gogo plugin
This was located in protobuf/plugin, which is imported by swarmkit code.
Because it imports some extra protobuf packages, it was adding
dependencies to code that vendors swarmkit.

Move this code to cmd/protoc-gen-gogoswarm, since it's only needed for
this code generation tool. This avoids new dependencies for swarmkit
downstreams.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>