allocator: Fix stuck tasks at startup
If the allocator starts up and sees a task in the NEW state that doesn't
need allocations, it should move it to PENDING immediately.

There was code in place for this but it wasn't actually updating the
task's state. The task would get stuck and never get assigned.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit 1c6c9f1beb10701ce92f4a6e230981931acd6275)