Initial opentracing implementation
This commit adds opentracing to swarmkit's manager and nodes, allowing
distributed tracing from service creation on a swarm manager down to
task handling on each node.

So far tracing has been added only to replicated service creation.

The parent span produced via creating a service is encoded into the
service's annotations, which is propagated down to each task via the
task's ServiceAnnotations.

This allows all future functionality (eg. when the scheduler runs
`applySchedulingDecisions` on an individual task, or when a node transitions a
task from ASSIGNED to RUNNING) to use the encoded span as the parent,
allowing end to end tracing.

We've added two initial tracers to power tracing.  The first is Uber's
Jaeger tracer, which is zipkin compatible.  This allows any user to
create either a zipkin server or uber's Jaeger server to collect span
information and visualize the spans collected from each swarmd client.

The second is lightstep, which via an API key allows users to send all
span information to lightstep.com and visualize spans using their
dashboard.

Further implementation needs to occur within the dispatcher and global
orchestrators.

Note that this commit does not include dependencies in order to make
reviewing easier.

Signed-off-by: Tony Holdstock-Brown <tony@docker.com>