controlapi: Allow a spec to be provided for a server-side rollback
We recently added support for server-side rollbacks, where the
controlapi copies the previous spec over, instead of having the client
supply it. This lets the manager recognize this update as a rollback, so
it can honor the rollback-specific update parameters.

The problem with this is that it would break the existing ability to
initiate a rollback while changing certain parameters, such as
`--rollback --update-delay 0s` (or `--rollback --rollback-delay 0s`
once rollback parameters are supported). To avoid introducing a
regression, allow the client to optionally provide a spec while manually
initiating a rollback. This will override PreviousSpec.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>