updater: Decrease default monitoring period
The updater monitors a tasks after they start to make sure they don't
fail immediately. The default time interval for this is 30s. However,
this is a bit long for synchronous updates to wait. Reduce it to 5s.
Users who would like a longer monitoring period should specify their own
value, or ideally use a health check in the service definition.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>