Add support for automatic indexing over protobuf content
The user can supply a google.protobuf.Descriptor that serves as a schema
for their protobuf types. They can also specify a set of fields that
they want to index.

The store will first compile each requested index into a path of fields,
for example "4, 2, 9" to indicate the 9th field inside the message
embedded in the 2nd field inside the message embedded inside the 4th
field of the top-level object.

Then, when doing actual indexing, it doesn't need to do any expensive
name lookups. It only parses the marshalled protobuf content to extract
the features specified by the user.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>