Fix possible race in ListenSubscriptions
It was possible, if a node left at the same time that a new subscription
for that node was added, that the subscription could be left un-Done and
the logs might hang. This change indexes nodes->subscriptions and
cancels a node's subscriptions when it leaves under the protection of a
mutex, which guarantees no subscription can be left open.

Also, wrap the subscription's Done method with a method on LB so that we
can correctly remove subscriptions from a node's subscription list when
a sub is marked as done.

Signed-off-by: Drew Erny <drew.erny@docker.com>