scheduler: Tasks with a desired state < RUNNING should count
Currently, the scheduler counts tasks on each node with desired state of
RUNNING. This works fine for sequential updates, but in parallel
updates, the scheduler can assign tasks unfairly, because the ones it
already assigned may not have changed their desired state to RUNNING
yet. In an update with unlimited parallelism, all the tasks may end up
on the same node, even when there are multiple nodes that can accomodate
the tasks.

Fix this by considering all tasks with a desired state at or below
RUNNING to be active. For example, tasks at READY should be seen the
same way as RUNNING tasks by the scheduler - they were assigned to that
node and will eventually start running.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>