node: Try to force TLS certificate renewal if role doesn't change after demotion
Currently, we wait 16 seconds for the role to change, and restart the
manager if it doesn't happen. This may be the right thing to do, because
if the node is promoted again, we might never see a role change. But
it's also possible that the 16 second timeout can be hit even with no
promotion, and this is a bad failure mode because it unexpectedly
re-adds the node to the Raft cluster.

Address this by being extra careful when the node was demoted, and
trying to force a certificate renewal in the case where we still don't
see a role change after 16 seconds.

This also adds RenewTLSConfig to node's waitgroup, to avoid some test
failures where RemoveAll returned an error because RenewTLSConfig was
still running.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>