integration: Improve logging
Since integration tests tests run in parallel, it can be very hard to
tell which test a log line came from.

Add a "testname" field to the logger used for the tests.

Plumb context through roleManager so that its log messages will benefit
from this.

Also preserve the logger in (*Node).run and (*Node).runManager so these
injected fields reach further.

I considered turning on debug logging, but it's an enormous amount of
information.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>