Fixed logs hanging on preparing tasks
When we're not following logs, we don't want any tasks that aren't
running yet; they don't have any logs to give us at the time of the
call. This is problematic because when a worker gets a subscription, if
that subscription includes a task that isn't running yet, it will wait
until the task is running to return logs. A task may be stuck in the
preparing state if an image takes a long time to pull, if, for instance,
if Amazon S3 were to go down.

Changes subscription to pass over adding tasks that aren't
running yet, and to add an error message explaining logs may be
incomplete.

Signed-off-by: Drew Erny <drew.erny@docker.com>