all: Import crypto/rand as "cryptorand"
I'm a bit uncomfortable with importing crypto/rand as "rand", because
this name can clash with math/rand. The packages both define a Read
function. If code is moved around and an IDE helpfully imports math/rand
for something that needs cryptographic randomness, this can introduce a
big vulnerability.

Import crypto/rand as "cryptorand" instead to make this less ambiguous.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>