node: Consume notifyNodeChange while requesting certificate renewal
Sometimes node shutdown can hang in agent.Leave. It turns out that the
agent main loop is stuck in handleSessionMessage, trying to write to
notifyNodeChange. The goroutine that reads from this channel may be
blocked in forceCertRenewal if the renewal process can't keep up (maybe
the manager has already shut down, and we're retrying several times).
Continue to consume from notifyNodeChange while trying to write to
forceCertRenewal.

Also, this goroutine should not exit until the agent is finished. Having
it exit when the context is cancelled may cause the deadlock if the
manager exits first, and the agent is stuck trying to write to
notifyNodeChange.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>