controlapi: Allow removal of network when only dead tasks reference it
RemoveNetwork currently won't allow a network to be removed if any
task in the store references it. This is a change relative to Docker
1.12, where only services were checked.

This might be a little heavy handed. Some tasks in the store may be long
dead. Resource attachment tasks, in particular, must be detached by the
API user, and this might be neglected after a failure.

Change RemoveNetwork to only fail the network deletion if a task has
actual and desired state of "running" or below. This means that neither
a failed resource attachment task (whose desired state won't be
adjusted) or a down node (where actual state won't reflect desired
state) can block the network deletion.

There may be some corner cases where deleting the network while a task
is in the process of shutting down prevents the allocator from
deallocating that task, but on balance this seems better than allowing
many situations to block network deletion indefinitely.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit b00598c99eeb755f5922ccd0a311f540df4faa3c)