dispatcher: Don't look up secrets by ID prefix
The SecretID field of a SecretReference must contain the full ID of the
secret. controlapi will reject the SecretReference if it contains
something other than a valid secret ID.

The dispatcher can look up secrets using GetSecret instead of trying to
look them up by prefix. This simplifies the code and makes it clearer
that the prefix matching functionality is not applicable here.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>