dispatcher: Fix possible format string bug in logging
There are a few places where Debugf/Errorf are used in places where a
format string is not expected. In most instances this is cosmetic, but
the case of `log.G(ctx).Debugf(err.Error())` is problematic. If the
error message contains a % character, the output will be malformed.
Correct these log calls.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>