dispatcher: Simplify Assignments
Rather than having separate maps for updates and removals of each type
of resource, have a single unified map that tracks all changes. This
makes it simpler to assemble the update sent to the agent: the last
change for any given ID always wins, whether it's a creation, update, or
removal.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>