Version service specs
Adds fields to Service and Task that keep track of a version number for
ServiceSpecs. This version number is different from the service object
version (which is tied to the Raft index).

Then change the scheduler to use this, instead of marshalling the specs
to compare them (which didn't work reliably).

Also make the orchestrator use SpecVersion as an optimization, when
available.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>