store: Fix type wrappers for Resource and Extension types
The Resource and Extension types still use wrappers around the protobuf
type. This will be necessary for future changes that reference the
relevant extension inside the resourceEntry, and reference indexer
objects in the extensionEntry. The use of wrappers isn't working
correctly in the current state, because CopyStoreObject is returning the
base protobuf type, and stripping the wrapper before anything is saved
in the store. Fix this by defining a custom CopyStoreObject for each of
these types. Also, wrap the indexers.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>