controlapi: Allow network attachments to be changed
We disallow changing network attachments, because in the past this
wouldn't trigger rolling updates properly. However, this restriction
predates the movement of Networks from Spec to TaskSpec. It should be
fine to update TaskSpec.Networks, since the criteria for a rolling
update is the TaskSpec in the task not matching the service.

Change controlapi to allow the Networks field in TaskSpec to be updated.
It's also okay to change the field in Spec if the one in TaskSpec is
also being changed at the same time - for example, if migrating data
from the legacy Spec.Networks field to Spec.Task.Networks.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>