raft: Give up leadership if the store is wedged
There have been a few bugs in the past that involved deadlocks where the
store lock was held indefinitely. These bugs are particularly bad
because the manager continues to send heartbeats as normal, so there's
no opportunity for a leader election to replace the stuck leader.

Add a method to MemoryStore that returns true if the lock has been held
for more than 30 second. Check this method in the raft implementation,
and use TransferLeadership when the store gets stuck.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>