Regenerate protobuf bindings with go1.8
This causes changes because of
https://github.com/golang/go/commit/b4e714e59efd21f9140a3e8e2b0182c74bfde437
which causes the mtime embedded in the compressed descriptor data to go from
invalid (Mon 14 Jul 02:04:00 BST 2042, which modulo timezones is the date in
the commit) to zero.

This means that anyone running go<1.8 will see the inverse change and fall foul
of `make checkprotos` during CI etc.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>