dispatcher: Clean up leftover task and node updates
The leader sets timers that trigger when nodes' heartbeat timeouts
expire. These timers may trigger after another node takes over as the
leader. In that case, they add updates to nodeUpdates that are acted on
when this node becomes the leader again. That's bad because it means
other nodes will immediately be marked as "DOWN" instead of being in the
UNKNOWN state for the ~15 second grace period.

Reset nodeUpdates and taskUpdates when the dispatcher starts up, so
stale data is not a factor.

I found I could reproduce this reliably by running a cluster of three
managers, and using ^Z to suspend the leader. After repeating this a few
times, nodes would start to get marked DOWN immediately after a leader
change.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>