raft: Don't log errors removing members based on snapshot
When we remove the peers on the snapshot's Removed list from from the
transport, we may have already removed them in the past, or never even
been in contact with these peers. We shouldn't log an error if it turns
out these peers are not known by the transport. This avoids a misleading
error message.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>