allocator: Correct batch logic
If an update to Raft did not succeed, the wrong object would potentially
be removed from the unallocated list. The logic that cleaned up the
unallocated list assumed that all updates to Raft succeeded, which is a
bad assumption to make.

This shouldn't have been a problem in practice, because there are only
two errors we would expect to see. One is ErrSequenceConflict, which is
handled in a way that shouldn't result in a second ErrSequenceConflict.
(Note that this way of overriding sequence conflicts is problematic, but
fixing it would be a somewhat large/risky change). The other possible
error is loss of leadership, in which case the allocator will be shut
down and retries become moot.

Even though the current code doesn't seem likely to lead to problems,
it's worth fixing the bogosity.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>