store: Simplify Batch
Currently Batch returns the number of updates it was able to commit
successfully, but this isn't really actionable. Either it committed
everything successfully, or the node lost leadership and it doesn't
matter what was committed and what wasn't (the new leader will pick up
from where we left off). A compliant caller of Batch would have to do a
bunch of accounting keep its internal state in sync with what Batch was
able to commit. This turns out not to matter in the current
implementation. Most uses of Batch ignored the first return value.

Change Batch to only return an error.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>