node: Support older CAs which don't set Node.Role
The Node.Role field was added in 17.04. Managers running earlier
versions don't set this field, which triggers a certificate renewal loop
in nodes running 17.04 or later.

To fix this, trigger certificate renewals off the DesiredRole field as
well. This may result in an extra certificate rotation when interacting
with a current CA, since the certificate will first be renewed when
DesiredRole changes, then again when Role changes.

Also, make rotations edge-triggered based on changes in the Role and
DesiredRole fields, to avoid rotation happening repeatedly in a loop.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>