Actually enable CheckQuorum
A change to enable the CheckQuorum option was merged in September, but
never actually took effect because the config structure where it was set
is not used directly.

This fixes the problem to enable it.

The option is useful because it makes a leader realize that quorum is
lost and start a new election. Otherwise, the leader will still try to
service RPCs, and will succeed when those RPCs don't involve writes.
This is confusing.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>