Support templated secrets and configs
Add the ability for secrets and configs to expand templates, which can
reference the following fields:

- The same set of fields available to ContainerSpec templates
  ({{.Service.Name}}, {{.Node.ID}}, etc)
- Environment variables: {{Env "envvar"}}
- Other secrets: {{Secret "sometarget"}}
- Other configs: {{Config "sometarget"}}

Secrets and configs can be accessed either by source (secret name) or
target (file name).

Templating of secrets/configs is optional, and is disabled by default
because it's very possible for Go template syntax to conflict with the
syntax of the payload.

The only change that will be necessary in the executor is wrapping the
DependencyGetter in NewTemplatedDependencyGetter before handing it off
to the execution backend.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>