Avoid writing to Network.Spec in allocatePools
A NetworkSpec should never be updated by the core. Avoid updating it in even a
trivial manner by restructuring the creation of ipamConfigs. I also happen to
find the new logic easier to grok.

This is largely benign right now but never writing to Network.Spec will make
changes in #2108 easier to reason about.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>