agent: avoid race condition after failing task startup
This moves the publish of a task manager creation event to after the
successful resolution of a task manager, ensuring that it will be there
when later resolved for log subscriptions.

Before, we correctly ignored the error, but it was not reported via the
correct log messagses. We have added the proper `WithError` to the log
message to ensure traceability.

It is unclear whether or not this early mount validation, the cause of
the controller resolution failure, is valid or not. We should really be
deferring to the "runtime" to make that determination and fail at
startup, but this PR should allow this to work gracefully.

Signed-off-by: Stephen J Day <stephen.day@docker.com>