node: Fix cert renewal logic race
It's possible to receive the current cert just after making the decision
of whether the cert needs to be renewed. In a rapid promote/demote
cycle, the node might decide that it doesn't need to renew its cert for
a role change to "manager" because it's already a manager, but then just
afterwards receive notice that its cert role is now "worker". With the
current logic, it gets stuck in this state and never triggers a renewal
to become a manager again.

This changes the logic to avoid comparisons to the current role, which
are inherently racy. Now, the cert will be renwed if the desired role
changed relative to the initial cert role or latest desired role, or if
the actual role changed and we were expecting that to happen based on
a previous change in desired role. The second half prevents us from
renewing every time we start up if we're talking to an old CA that
doesn't set Role.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>