node: Third proposed fix for cert renewal logic race
This is similar in concept to the last attempt, but moves the retries
into the ca code. RenewTLSConfig had to be refactored to support
notifying the loop of the expected role. This couldn't be passed over
the channel easily, because the sending code can't block indefinitely
long, and using a goroutine would let updates to the expected role race.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>