Temporarily vendor fork of etcd
Now that we are using CheckQuorum, it is causing some spurious leader
elections that are leading to test failures in both SwarmKit and Docker
integration tests.

I submitted this change upstream, but it hasn't been reviewed yet. I'd
like to temporarily vendor a fork that has the change. First, this will
unblock vendoring swarmkit from the master branch. Also, it will give us
a chance to test the change more broadly. Finally, we probably don't
want to vendor the master branch of etcd, so without this temporary
fork, we'd have to wait for a release that includes this change.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>