agent: Resolve deadlock in run
updateNode uses sendError to force a new session. The problem is that
sendError sends to a channel that is read by the loop which calls
updateNode. The channel only has a buffer of 1. If multiple errors
happen at the same time, this can cause a deadlock. One example of when
this can happen is if nodeUpdateTicker fires, causing the session to be
closed, and then immediately after, the select chooses
NotifyTLSChange, which tries to close the session a second time but
blocks.

Have updateNode terminate the session directly, instead of sending over
a channel to a receiver that's in the same goroutine.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>