dispatcher: Quiet log messages when a service is deleted
Deleting a service causes the underlying tasks to be deleted. The
dispatcher currently logs warning and error messages when it can't find
tasks that correspond to status reports from an agent. These log
messages should be logged at the debug level, because it's normal for
tasks to be deleted like this. Also, only one message should be logged
per missing task.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>