raft: Check for leader before returning quorum loss error
If the leader comes back after an extended period of quorum loss, we may
not realize this until a tick has elapsed. API calls may fail even
though there actually is a leader at that point in time. To solve this,
optimistically try to get the leader connection, and only check
ticksWithNoLeader if this fails.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>