Only filter on plugins if running with docker engine.
If there is no engine description then thus assumes that all drivers etc are
built-in and this always available, and when they are not proper errors will be
generated and propagated.

This avoids the need in https://github.com/docker/swarmkit/pull/1965 to
populate a spurious `api.NodeDescription.Engine` field (spurious because there
is no engine in this case).

Signed-off-by: Ian Campbell <ian.campbell@docker.com>