Convert NetworkAllocator to an interface
... and move the current implementation to cnmallocator.

I would like to implement support for CNI to be used with the containerd
executor (#1965), in my prototypes I ended up scatering lots of `if
a.networkModel == api.NetworkModel_{CNI,CNM}` throughout
`manager/allocator/network.go` and elsewhere. By making this an `interface`
instead I will be able to add a cni implementation (which should be quite
trivial) instead.

The networkallocator code exposed a few things directly to outside packages
(i.e. not directed through the allocator interfaces), rework those. In some
cases this required plumbing through an allocator to be used.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>