scheduler: Fix flaky test TestSchedulerCompatiblePlatform
task1 lands on node1.
task2 cannot be scheduled.
task3 can land on either node2 or node3 without creating an unbalanced
situation. The lack of platform information on node3 is not a factor
because task3 doesn't have platform constraints.

However, the test currently expects task3 to land on node1 or node2.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>