Abstract Network Model into an interface
Currently the implementation is the existing CNM provider.

Since plugingetter is in effect CNM specific encapsulate it into the CNM
implementation of network.Model instead of passing it around. This means moving
validate driver (which was only used for networking purposes) down into the
interface too. This also allows moves the caller of
allocator.IsBuiltInNetworkDriver to the CNM model, which can therefore call the
cnmallocator directly.

One downside of this approach is that while the plugingetter was almost always
optional the network model is not, meaning one needs to be added to numerous
test cases.

The uses of plugingetter are now confined to the CNM model and allocator code
rather than being exposed to various parts of the manager and node, which seems
cleaner.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>