allocator: Fix events watch leak
The allocator starts a watch on events and never shuts it down. If this
node loses leadership, events will continue to pile up in that queue
forever, leading to memory exhaustion.

Fix this by giving the watch a well-defined lifecycle. Return it along
with a cancel function from an init() method to make it clear it must be
cancelled, instead of sticking the cancel function value in a struct
where there is no clear responsibilty for calling it.

In the future, we might consider changing Watch to run a callback
function, so shutting down the watch is mandatory (similar to what we do
with store transactions).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit c21d70e4cf2536ee37b7f56c745cb9818d679123)