taskreaper: Clean up Watch code
It is not necessary to create a watch in the New function. Creating it
in Run is more consistent with other parts of the code, and makes the
lifecycle of the watch clearer.

This is not a functional change, but it may help prevent adding bugs in
the future.

Also, thread through a context to use for logging.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>